requirements.txt
PyMuPDF>=1.24.0
flask>=3.0
anthropic>=0.40.0
python-dotenv>=1.0
